add burnt
