+++ src/components/Toast/Toaster.tsx
@@ -1,0 +1,3 @@
+export function Toaster() {
+  return null
+}
