+++ src/components/Toast/index.e2e.ts
@@ -1,0 +1,5 @@
+import {ToastProps} from './index.types'
+
+export * as Icon from './ToastIcon'
+
+export function toast(_: ToastProps) {}
