+++ src/components/Toast/index.types.ts
@@ -1,0 +1,6 @@
+import {IconName} from './ToastIcon'
+
+export interface ToastProps {
+  message: string
+  icon?: IconName
+}
