--- src/view/com/util/Toast.e2e.tsx
@@ -1,1 +1,0 @@
-export function show() {}
